aUTH

+++ UserServiceApi/src/main/java/com/ecors/api/users/entity/Address.java
@@ -14,7 +14,7 @@ public class Address {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long addressId;
+	private Integer addressId;
 
 	@ManyToOne
 	@JoinColumn(name = "id")
@@ -31,13 +31,11 @@ public class Address {
 	private AddressType type;
 	private Boolean deliveryAddress;
 
-	
-
-	public Long getAddressId() {
+	public Integer getAddressId() {
 		return addressId;
 	}
 
-	public void setAddressId(Long addressId) {
+	public void setAddressId(Integer addressId) {
 		this.addressId = addressId;
 	}
